Fix template error handling; add logging support
The exceptions were being caught in the wrong order such that
`ConnectionError` exceptions, which are a sub-class of `TransportError`
were not being considered.

We also add support for only updating a template when the template's
mapping version is different.

We add a `logger` argument for `streaming_bulk()` so that we have a
bit more insight into what is happening while processing actions.

Further, we fix the failing unit tests, add Travis CI integration,
and ensure the code passes flake8.